Update: refactor projects file structure
Add IsBuilder permission
